Some improvements accumulated over last two weeks:
* PsychPortAudio: Some compile/link fixes for OS/X, prototype implementation for
  audio capture (non-duplex up to now), unfinished.

  We need to ship the portaudio.0.019.dylib for OS/X and the user needs to
  manually install it :-(. Its not possible to statically link PortAudio
  and PsychPortAudio due to some weird restrictions of the OS/X dynamic
  linker with respect to mulitply defined symbols in static libraries.

* Alpha version of latency/timing test signal generator for PsychPortAudio:
  PsychPortAudioTimingTest.m

* Some new features for the imaging pipeline: Builtin-function for use in hook slots
  to restrict image processing to some ROI via glScissor() scissor box testing.

  PsychImaging.m provices an easy to use high-level interface to this.

* New built-in, but untested imaging blitter for non-rectangular blits via
  use of GL display lists. This is the first bit of code for display
  geometry correction.

-> Screen for IntelMac updated.

* Thanks to Adam Greenberg: Documentation for DaqAIn and DaqAInScan improved,
  sign handling in DaqAIn fixed (some tests by Adam pending).

* New improved implementaiton of Ask.m by Adam Greenberg. Allows to select
  fontsize and provides builtin 'GetChar' method of response collection,
  similar to the old GetEchoString.m

* Some demos updated.

* BrightSide driver updated to the current implementation of imaging pipeline.
  Tests on our display pending (mex low-level driver needs recompile).




git-svn-id: https://psychtoolbox-3.googlecode.com/svn/trunk@790 f5efbe48-b203-0410-9429-fc5e098e5a54

+++ PsychSourceGL/Source/Common/Screen/PsychWindowSupport.c
@@ -1427,7 +1427,7 @@ double PsychFlipWindowBuffers(PsychWindowRecordType *windowRecord, int multiflip
 				// we've already verified correct working of the queries during startup.
 				if ((*beamPosAtFlip < 0) || (*beamPosAtFlip > vbl_endline)) {
 					// Ok, this is completely foo-bared.
-					printf("PTB-ERROR: Beamposition query after flip returned the *impossible* value %i (Valid would be between zero and %i)!!!\n", *beamPosAtFlip, vbl_endline);
+					printf("PTB-ERROR: Beamposition query after flip returned the *impossible* value %i (Valid would be between zero and %i)!!!\n", *beamPosAtFlip, (int) vbl_endline);
 					printf("PTB-ERROR: This is a severe malfunction, indicating a bug in your graphics driver. Will disable beamposition queries from now on.\n");
 					printf("PTB-ERROR: Timestamps returned by Flip will be correct, but less robust and accurate than they would be with working beamposition queries.\n");
 					printf("PTB-ERROR: It's strongly recommended to update your graphics driver and optionally file a bug report to your vendor if that doesn't help.\n");
